test: separate preflight tests
